requirements.txt
requests>=2.31.0
urllib3>=2.0.7
typing-extensions>=4.8.0
python-dotenv>=1.0.0
Appium-Python-Client>=3.1.1
selenium>=4.18.1
pyautogui>=0.9.54
python-xlib>=0.33
pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.4
redis>=5.0.1 